Allow autofinding of themes.
A first step in making adding themes easier.
